cleaned up the states in my project
created the base for my test for RegistrationViewModel
