+++ app/src/main/java/com/example/googlelightcalendar/core/viewmodels/login/LoginViewModel.kt
@@ -8,7 +8,6 @@ import com.example.googlelightcalendar.common.Constants
 import com.example.googlelightcalendar.interfaces.AppAuthClient
 import com.example.googlelightcalendar.navigation.components.NavigationBuilder
 import com.example.googlelightcalendar.navigation.components.NavigationDestinations
-import com.example.googlelightcalendar.navigation.components.NavigationDestinations.registerScreen
 import com.example.googlelightcalendar.navigation.components.NavigationManger
 import com.example.googlelightcalendar.repo.AuthorizationResponseStates
 import com.example.googlelightcalendar.repo.UserRepository
