--- app/src/main/java/com/example/googlelightcalendar/screens/loginScreen/LoginScreenStates.kt
@@ -1,1 +1,0 @@
-package com.example.googlelightcalendar.screens.loginScreen
