+++ app/src/test/java/com/example/googlelightcalendar/fakes/RegistrtrationCacheFake.kt
@@ -1,0 +1,4 @@
+package com.example.googlelightcalendar.fakes
+
+class RegistrtrationCacheFake {
+}
